feat: reduce the complexity of Client
WIP: The current Client works but is problematic if not polled.
In scenarios such as a select! loop another very noisy future can effectively 'drown out' Client, resulting in a timeout at the LFS end.
By drastically simplifying the client it will also allow other parties to engage with the project down the line.