refactor!: Object ergonomics, address issues with coordinate system: normalising vs not.
Attempts to addresses the fundamental issue of attempting to use glam directly - we
have an issue with "what units are we actually working with here, what
scale".

Attempted to normalise, however because of the precision issues between
the 1:65536 and 1:16 scaling systems you result in swapping one problem
for another. So it's better to recognise that actually they're
fundamentally different types, but provide gateway functions to working
in human values (metres).