feat!: insim v10 + Proper Object Handling + Updated PTH and new PIN file handling + other breaking changes (#266)
Updated packets for insim v10, this includes breaking changes from upstream
Removed support for legacy uncompressed packet length moving forward (codec::Mode and associated supporting functions on builder)
Change of default behaviour to request the protocol version from LFS
Alter encoding and decoding derive macros to remove unnecessary support for centiseconds, etc., now that all insim packets are milliseconds. This is a breaking change.
Updated tracks for LFS 0.8B
Refactored ObjectInfo to strongly type objects according to 0.8. This is a breaking change.
Added PIN file support. Support for multiple revisions of PTH file. This is a breaking change.
Remove insim_smx as SMX's are basically a dead format.
`strip` not returning a borrowed Cow when it should, vehicle de-duplication
Add ColourifyExt to allow chaining of colours for String type, as an alternative to format!
Revert Speed to something more akin to Duration from the std library. This is a breaking change. The idea behind SpeedKind was useful in theory, but in practise it was just getting in the way. Humans almost always work in m/s, mph or kmph.
Rename Direction to Heading. This is a breaking change.,
Attempts to address the fundamental issue of attempting to use glam directly - we have an issue with "what units are we actually working with here, what scale". Because of the precision issues between the 1:65536 and 1:16 scaling systems you result in swapping one problem for another. So it's better to recognise that actually they're fundamentally different types, but provide gateway functions to working in human values (metres).